+++ Foundation/System/IO/File.h
@@ -27,25 +27,25 @@ namespace sys
 		~File();
 	public:
 		// 将数据写到本地
-		static bool write(const std::string& url, const char* data, int64_t size, int64_t& writtenSize);
+		static bool write(const std::string& url, const char* data, size_t size, size_t& writtenSize);
 		// 从本地读取数据
-		static char* read(const std::string& url, int64_t& readSize);
+		static char* read(const std::string& url, size_t& readSize);
 		// 从本地读取数据
 		static bool read(const std::string& url, std::string& data);
 		// 追加数据到本地
-		static bool append(const std::string& url, const char* data, int64_t size, int64_t& appendSize);
+		static bool append(const std::string& url, const char* data, size_t size, size_t& appendSize);
 		// 判断文件是否存在
 		static bool exists(const std::string& url);
 	public:
 		bool open();
 		// 将数据写到本地
-		bool write(const char* data, int64_t size, int64_t& writtenSize);
+		bool write(const char* data, size_t size, size_t& writtenSize);
 		// 从本地读取数据
-		char* read(int64_t& size);
+		char* read(size_t& size);
 		// 从本地读取数据
 		bool read(std::string& data);
 		// 追加数据到本地
-		bool append(const char* data, int64_t size);
+		bool append(const char* data, size_t size);
 		// 判断文件是否存在
 		bool exists();
 		// 关闭
