+++ Foundation/System/Stream/FileStream.cpp
@@ -29,7 +29,7 @@ void FileStream::load(const std::string& filepath)
 		return;
 	}
 
-	int64_t size = 0;
+	size_t size = 0;
 	char* data = File::read(filepath, size);
 	if (data == nullptr)
 	{
@@ -45,7 +45,7 @@ void FileStream::save(const std::string& filepath)
 	{
 		return;
 	}
-	int64_t writtenSize = 0;
+	size_t writtenSize = 0;
 	File::write(filepath, this->getData(), this->getLength(), writtenSize);
 }
